线程池的测试

+++ server/Socket.h
@@ -104,7 +104,8 @@ public:
 	virtual int Close() {
 		m_status = 3;
 		if (m_socket != -1) {
-			unlink(m_param.ip);
+			if (m_param.attr & SOCK_ISSERVER)
+				unlink(m_param.ip);
 			int fd = m_socket;
 			m_socket = -1;
 			close(fd);
